Move duplicated utility functions into linker-utils
